# db: keep up to one memtable for recycling